+++ administrator/components/com_menus/presets/alternate.xml
@@ -467,7 +467,7 @@
 				title="MOD_MENU_COM_TEMPLATES_SUBMENU_TEMPLATES"
 				type="component"
 				element="com_templates"
-				link="index.php?option=com_templates&amp;view=styles&amp;client_id=0"
+				link="index.php?option=com_templates&amp;view=templates&amp;client_id=0"
 			/>
 
 			<menuitem
