--- administrator/components/com_templates/tmpl/styles/default.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<metadata>
-	<layout title="COM_TEMPLATES_STYLE_VIEW_DEFAULT_TITLE">
-		<message>
-			<![CDATA[COM_TEMPLATES_STYLE_VIEW_DEFAULT_DESC]]>
-		</message>
-	</layout>
-</metadata>
