+++ administrator/components/com_templates/tmpl/templates/default.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<metadata>
+	<layout title="COM_TEMPLATES_TEMPLATES_VIEW_DEFAULT_TITLE">
+		<message>
+			<![CDATA[COM_TEMPLATES_TEMPLATES_VIEW_DEFAULT_DESC]]>
+		</message>
+	</layout>
+</metadata>
