+++ administrator/language/en-GB/mod_menu.ini
@@ -22,7 +22,6 @@ MOD_MENU_COM_PRIVACY_CAPABILITIES="Capabilities"
 MOD_MENU_COM_PRIVACY_CONSENTS="Consents"
 MOD_MENU_COM_PRIVACY_DASHBOARD="Dashboard"
 MOD_MENU_COM_PRIVACY_REQUESTS="Requests"
-MOD_MENU_COM_TEMPLATES_SUBMENU_STYLES="Styles"
 MOD_MENU_COM_TEMPLATES_SUBMENU_TEMPLATES="Templates"
 MOD_MENU_COMPONENTS="Components"
 MOD_MENU_CONFIGURATION="Global Configuration"
